requirements.txt
python-telegram-bot >= 13.13
google-api-python-client >= 2.58.0
google-auth-httplib2 >= 0.1.0
google-auth-oauthlib >= 0.5.2
selenium==4.4.3